circle line color

+++ src/parser/NodeCircle.java
@@ -23,7 +23,7 @@ public class NodeCircle extends ResponsibleChainNode {
     @Override
     protected void exec(GeomCompos geom, ArrayList<Float> args) {
         Circle circle = new Circle(new Points(args.get(0),args.get(1)),args.get(2));
-        circle.setColor(new Color(args.get(3),args.get(4),args.get(5)));
+        circle.setColor(new Color(args.get(3).intValue(),args.get(4).intValue(),args.get(5).intValue()));
         geom.addGeometric(circle);
     }
 }
