+++ src/parser/NodeLine.java
@@ -20,7 +20,7 @@ public class NodeLine extends ResponsibleChainNode{
         Points p1 = new Points(args.get(0),args.get(1));
         Points p2 = new Points(args.get(2),args.get(3));
         Line line = new Line(p1,p2);
-        line.setColor(new Color(args.get(4),args.get(5),args.get(6)));
+        line.setColor(new Color(args.get(4).intValue(),args.get(5).intValue(),args.get(6).intValue()));
         geom.addGeometric(line);
     }
 }
